added alert to signout

+++ snap_n_claim/lib/screens/employee/employee_menu_drawer.dart
@@ -80,14 +80,36 @@ class EmployeeMenuDrawer extends StatelessWidget {
 
   //handle signout
   Future<void> _onTapSignOutBtn(BuildContext context) async {
-    final directory = await getApplicationDocumentsDirectory();
-    final path = directory.path;
-    File file = File('$path/userdata.txt');
-    await file.delete();
+    var dialogRes = await showDialog<bool>(
+        context: context,
+        builder: (context) => AlertDialog(
+          actionsAlignment: MainAxisAlignment.spaceBetween,
+          content:
+          const Text('Are you sure you want to signout?'),
+          actions: <Widget>[
+            TextButton(
+              onPressed: () => Navigator.pop(context, false),
+              child: const Text('No'),
+            ),
+            TextButton(
+              onPressed: () async {
+                return Navigator.pop(context, true);
+              },
+              child: const Text('Yes'),
+            ),
+          ],
+        ));
 
-    Navigator.of(context).pop();
-    Navigator.of(context).pushReplacement(MaterialPageRoute(
-        builder: (BuildContext context) => LoginScreen(_width, _height)));
+    if(dialogRes == true){
+      final directory = await getApplicationDocumentsDirectory();
+      final path = directory.path;
+      File file = File('$path/userdata.txt');
+      await file.delete();
+
+      Navigator.of(context).pop();
+      Navigator.of(context).pushReplacement(MaterialPageRoute(
+          builder: (BuildContext context) => LoginScreen(_width, _height)));
+    }
   }
 
   @override
